requirements.txt
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==24.8.1
beautifulsoup4==4.13.3
certifi==2025.1.31
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.1
constantly==23.10.4
cryptography==44.0.2
daphne==4.1.2
defusedxml==0.7.1
Django==5.2
django-debug-toolbar==5.1.0
django-extensions==4.0
django-ranged-response==0.2.0
django-simple-captcha==0.6.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
flake8==7.2.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
isort==6.0.1
mccabe==0.7.0
msgpack==1.1.0
oauthlib==3.2.2
packaging==24.2
pillow==11.1.0
psycopg==3.2.6
psycopg-binary==3.2.6
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.13.0
pycparser==2.22
pyflakes==3.3.2
PyJWT==2.9.0
pyOpenSSL==25.0.0
python-dotenv==1.1.0
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
service-identity==24.2.0
setuptools==78.1.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
soupsieve==2.6
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.13.1
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
zope.interface==7.2
